splash/requirements.txt
git+https://github.com/huggingface/transformers.git
torch
torchaudio
soundfile
librosa
moviepy
matplotlib
pillow
